Generate shared attention
